DATACOUCH-169 - Automatic index creation can be toggled via bean
IndexManager now can be constructed with ignore parameters to ignore each type of index. The instance used in the repository factory is created as a Bean. This allows for changing the behavior of automatic index creation at runtime depending on the Spring context (eg. deactive it only in Prod through a Profile-annotated specific configuration).

Removed the target METHOD from the N1qlSecondaryIndex since the fields cannot be customized.